Organized the code using Packages
All GUI code will be in GraphicalUserInterface Package

Backend will be in BackendCode Package

Command Line will be in Command Line Interface
