+++ UML editor/src/main/java/codecain/BackendCode/SaveManager.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
