+++ UML editor/src/main/java/codecain/BackendCode/UMLClass.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 import java.util.HashMap;
 import java.util.Map;
