+++ UML editor/src/main/java/codecain/BackendCode/UMLClassInfo.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 import java.util.ArrayList;
 import java.util.List;
