+++ UML editor/src/main/java/codecain/BackendCode/UMLFields.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 /**
  * The UMLFields class provides functionality to manage fields in a UML class.
