+++ UML editor/src/main/java/codecain/BackendCode/UMLMethods.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 import java.util.List;
