+++ UML editor/src/main/java/codecain/BackendCode/UMLParameterInfo.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.BackendCode;
 
 import java.util.Objects;
