+++ UML editor/src/main/java/codecain/CommandLineInterface/CLI.java
@@ -1,4 +1,4 @@
-package codecain;
+package codecain.CommandLineInterface;
 
 import javax.swing.*;
 import java.awt.*;
